% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCropWR.R
\name{calcCropWR}
\alias{calcCropWR}
\title{Crop Water Requirement}
\usage{
calcCropWR(
  data,
  Crop = NULL,
  pheno = c(Ph_ini = NULL, Ph_dev = NULL, Ph_mid = NULL, Ph_late = NULL),
  Kc.Type = "Single",
  Kc = c(Kc_ini = NULL, Kc.mid = NULL, Kc.end = NULL),
  CropHmax = NULL,
  plantDate = NULL,
  WHC = NULL,
  SoilTex = NULL
)
}
\arguments{
\item{data}{a dataframe containing the required climate data, see \code{\link{calcEto}}}

\item{Crop}{character. Name of the crop , Default: NULL}

\item{pheno}{length of crop growth stages, Default: c(Ph_ini = NULL, Ph_dev = NULL, Ph_mid = NULL, Ph_late = NULL)}

\item{Kc.Type}{the type of Kc used for estimation, either "single" or "double", Default: 'Single'}

\item{Kc}{defaoult Kc values in the form of: c(Kc_ini = NULL, Kc.mid = NULL, Kc.end = NULL)}

\item{CropHmax}{maximum crop height in (m)}

\item{plantDate}{a dataframe containing columens with 'Year' in YYYY and 'plantDate' in julian date.}

\item{WHC}{soil water holding capacity}

\item{SoilTex}{a character specfying the name of soil textural class according to USDA}
}
\value{
The function generates a data frame containing the following components:

\emph{\code{Crop: the name of the crop that water requirement is estimated.}}

\emph{\code{Eto: .daily estimations of reference crop evapotranspiration (mm/day).}}

\emph{\code{tabKc: daily tabulated Kc values over the growing period.}}

\emph{\code{adjKc: daily adjested Kc values over the growing period.}}

\emph{\code{CropWR: daily estimates of crop water requirement over the growing period.}}
}
\description{
This function computes crop water requirement/water demand and crop coefficients using the method described in FAO56
}
\examples{
# load example data:
data(AgroClimateData)

}
\references{
Allen, R.G.; Pereira, L.S.; Raes, D.; Smith, M. Crop Evapotranspiration: Guidelines for Computing Crop Water Requirements; FAO
Irrigation and Drainage Paper no. 56; FAO: Rome, Italy, 1998; ISBN 92-5-104219-5.
}
\seealso{
\code{\link{calcEto}, \link{calcSeasCal},  \link{calcWatBal}}
}
