% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSeasCal.R
\name{calcSeasCal}
\alias{calcSeasCal}
\title{Rainy Season Calandar}
\usage{
calcSeasCal(
  data,
  onsetWind.start,
  onsetWind.end,
  e_thresh = 0.25,
  AW_thr = 10,
  soilWHC
)
}
\arguments{
\item{data}{= a dataframe. It should bu an object as returned by calcWatBal}

\item{onsetWind.start}{onset start}

\item{onsetWind.end}{onset end}

\item{e_thresh}{thgreshold}

\item{AW_thr}{PAW}

\item{soilWHC}{Whater holding capacity of the soil}
}
\value{
The function generates a data frame containing the following components:

\code{Year:} { year }

\code{Onset.DOY:} { onset in DOY }

\code{Onset.index:} {onset index  }

\code{Cesation.DOY:} {  ce in DOY }

\code{Cesation.index:} { ce in index  }

\code{SeasDur:} {  durtion of the season }
}
\description{
This function calculates rainy season calander based on Agroclimatic approach
}
\examples{

data(climateData)

Eto.daily <- calcEto(climateData)

climateData$Eto <- Eto.daily$ET.Daily

 soilWHC = 100

watBal.daily <- calcWatBal(climateData, soilWHC)

onsetWind.start = "1980-09-01"
onsetWind.end = "1981-01-31"

seasCal.dF <- calcSeasCal(watBal.daily, onsetWind.start, onsetWind.end,
                          e_thresh = 0.25, AW_thr = 10, soilWHC)

}
\references{
Allen, R.G.; Pereira, L.S.; Raes, D.; Smith, M. Crop Evapotranspiration: Guidelines for Computing Crop Water Requirements; FAO Irrigation and Drainage Paper no. 56; FAO: Rome, Italy, 1998; ISBN 92-5-104219-5.

Doorenbos, J. and Pruitt, W.O. 1975. Guidelines for predicting crop water requirements, Irrigation and Drainage Paper 24, Food and Agriculture Organization of the United Nations, Rome, 179 p.
}
