% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWatBal.R
\name{calcWatBal}
\alias{calcWatBal}
\title{Soil Water Balance}
\usage{
calcWatBal(data, soilWHC)
}
\arguments{
\item{data}{= a dataframe containing the required climate variables: Columns must contain the following parameters:

\if{html}{\out{<div class="sourceCode">}}\preformatted{          Station_Name: weather station name
          Lat: latitude of the site in decimal degrees
          Lon: longitude of the site in decimal degrees
          Elev: elevation above sea level (m)
          Year: year in YYYY format
          Month: month in MM format
          Day: day of record
          Rain:
          Tmax: daily maximum temperature at 2m height (°C)
          Tmin: daily minimum temperature at 2m height (°C)
          Eto:
}\if{html}{\out{</div>}}}

\item{soilWHC}{Whater holding capacity of the soil}
}
\value{
The function generates a data frame containing the following components:

\code{cumRAIN:} {  }

\code{DEMAND:} {  }

\code{RUNOFF:} {  }

\code{ERATIO:} {  }

\code{AVAIL:} {   }
}
\description{
This function calculates a daily water balance computation for the root zone according to algorithms described in the FAO Irrigation
and drainage paper 56
}
\examples{

data(climateData)

Eto.daily <- calcEto(climateData)

climateData$Eto <- Eto.daily$ET.Daily

 soilWHC = 100

watBal.daily <- calcWatBal(climateData, soilWHC)

plot(watBal.daily$ERATIO*100, ty="l")
lines(watBal.daily$Eto, col="red")
lines(watBal.daily$Rain, col="blue")

}
\references{
Allen, R.G.; Pereira, L.S.; Raes, D.; Smith, M. Crop Evapotranspiration: Guidelines for Computing Crop Water Requirements; FAO Irrigation and Drainage Paper no. 56; FAO: Rome, Italy, 1998; ISBN 92-5-104219-5.

Doorenbos, J. and Pruitt, W.O. 1975. Guidelines for predicting crop water requirements, Irrigation and Drainage Paper 24, Food and Agriculture Organization of the United Nations, Rome, 179 p.
}
