% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcstWSC.R
\name{fcstWSC}
\alias{fcstWSC}
\title{Seasonal Forecast of WSC from Rainfall Probabilities}
\usage{
fcstWSC(sesRain, var.dF, rainTerc, variable)
}
\arguments{
\item{sesRain}{A data frame containing seasonal rainfall data with columns 'Year' and 'sRain'
(seasonal rainfall).}

\item{var.dF}{A data frame containing WSC variables including the year and the variable of
interest (e.g., 'onset' or 'cessation').}

\item{rainTerc}{A data frame with tercile probabilities for rainfall (columns 'T1', 'T2', and 'T3'
representing below normal, normal, and above normal terciles respectively).}

\item{variable}{A character string indicating the variable of interest ('onset' or 'cessation').}
}
\value{
A data frame containing the probabilities of the variable falling into the below normal,
normal, and above normal terciles (columns 'BN', 'NN', and 'AN' respectively).
}
\description{
This function forecasts the Wet Season Calendar (WSC) based on seasonal rainfall
totals and seasonal forecast of rainfall terciles.
}
