% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcn.R
\docType{data}
\name{rcn}
\alias{rcn}
\title{data which is used internally in the package}
\format{
An object of class \code{RasterLayer} of dimension 1809 x 2002 x 1.
}
\usage{
rcn
}
\description{
data which is used internally in the package
}
\keyword{datasets}
