% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEto.R
\name{calcEto}
\alias{calcEto}
\title{Potential Evapotranspiration}
\usage{
calcEto(data)
}
\arguments{
\item{data}{= a dataframe containing the required climate variables: Columns must contain the following parameters:

\if{html}{\out{<div class="sourceCode">}}\preformatted{          Station_Name: weather station name
          Lat: latitude of the site in decimal degrees [°]
          Lon: longitude of the site in decimal degrees [°]
          Elev: elevation above sea level [m]
          Year: year in YYYY format
          Month: month in MM format
          Day: day of record
          Tmax: daily maximum temperature at 2m height [°C]
          Tmin: daily minimum temperature at 2m height [°C]
}\if{html}{\out{</div>}}}
}
\value{
The function generates a list containing the following components:

\code{ET.Daily:} {Daily estimations of reference crop evapotranspiration (mm/day)}

\code{Ra.Daily:} {Daily estimations of extraterristrial radiation (MJ/m2/day)}

\code{Slope.Daily:} {Daily estimations of slope of vapour pressure curve (kPa/°C)}

\code{ET.type:} {Type of the estimation obtained}
}
\description{
This function calculates Penman-Monteith, Priestley Taylor and Hargreaves-Samani Potential Evapotranspiration
using the method described by Allen et al, (1998)
}
\examples{

calcEto(climateData)


}
\references{
Allen, R.G., L.S. Pereira, D. Raes, and M. Smith. 1998. ‘Crop evapotranspiration-Guidelines for Computing Crop Water requirements
}
