% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWatBal.R
\name{calcWatBal}
\alias{calcWatBal}
\title{Soil Water Balance}
\usage{
calcWatBal(data, soilWHC)
}
\arguments{
\item{data}{a dataframe containing the required variables: Columns must
contain the following parameters:

\if{html}{\out{<div class="sourceCode">}}\preformatted{   Lat: latitude of the site in decimal degrees.
   Lon: longitude of the site in decimal degrees.
   Elev: elevation above sea level in (meters).
   Year: year of record "YYYY".
   Month: month of record "MM".
   Day: day of record "DD".
   Rain: daily rainfall in (mm).
   Eto: daily potential evapotranspiration in (mm).
}\if{html}{\out{</div>}}}

\item{soilWHC}{\verb{  }Water holding capacity of the soil in (mm).}
}
\value{
The function generates a data frame containing the following
components:

\emph{\code{DRAIN: amount of deep drainage in (mm).}}

\emph{\code{TRAN: amount of water lost by transpiration in (mm).}}

\emph{\code{RUNOFF: surface runoff in (mm).}}

\emph{\code{AVAIL: available soil moisture storage in (mm).}}

\emph{\code{R: actual-to-potential evapotranspiration ratio.}}
}
\description{
Function to estimate a budget-based daily soil water balance.
It calculates the amount of water present in the root zone of a homogeneous
grass canopy growing on a well-drained and homogeneous soil.
}
\examples{
\donttest{
# load example data:
data(AgroClimateData)
}
}
\references{
Allen, R.G.; Pereira, L.S.; Raes, D.; Smith, M. Crop
Evapotranspiration: Guidelines for Computing Crop Water Requirements; FAO
Irrigation and Drainage Paper no. 56; FAO: Rome, Italy, 1998;
ISBN 92-5-104219-5.

Doorenbos, J. and Pruitt, W.O. 1975. Guidelines for predicting crop water
requirements, Irrigation and Drainage Paper 24, Food and Agriculture
Organization of the United Nations, Rome, 179 p.
}
\seealso{
\code{\link{calcEto}, \link{calcSeasCal}}
}
